requirements.txt
pytest==8.3.5
requests==2.32.3
pytest-mock==3.14.0
pytest-cov==6.1.1
numpy==2.2.6
PyYAML==6.0.2
hypothesis==6.131.21
pandas==2.2.3

